remove move system in favor of a motor component
