+++ src/ecs/components/Brain.component.js
@@ -1,3 +1,4 @@
+import _ from "lodash";
 import { Component } from "geotic";
 
 import { getEntity } from "../../lib/ecsHelpers";
@@ -5,76 +6,47 @@ import { getEntity } from "../../lib/ecsHelpers";
 import { world } from "../../ecs/index";
 
 export default class Brain extends Component {
-  // goals are entities with a goal component
-  // goal component is
-  // { isFinished<boolean>, takeAction<func>, originalIntent<goalEntityId> }
+  // todo: add basic instincts list as properties
+  // a jelly should have fewer basic instincts/desires than a dwarf
+  // for example, a dwarf may put a lot of weight to crafting things
+  // whereas a jelly has not concept of creativity at all
+  // the spirit could contain these desires
+  // these desires could change weights/orders based on external stim
+  // the spirit could possess something else - changing it's sense of self/behaviors
 
-  static properties = {
-    goalIds: [],
-  };
+  onTakeAction() {
+    // determine what type of goals to get
+    // const goalType = "BORED";
 
-  onDestroyed() {
-    this.goalIds.forEach((eid) => {
-      getEntity(eid).destroy();
-    });
-  }
-
-  onTakeAction(evt) {
-    while (
-      this.peekGoal() &&
-      this.peekGoal().goal.isFinished(this.peekGoal())
-    ) {
-      this.popGoal().entity.destroy();
-    }
-
-    const currentGoal = this.peekGoal();
-
-    if (currentGoal) {
-      currentGoal.goal.takeAction(currentGoal);
-      this.removeGoal(currentGoal);
-    }
-
-    evt.handle();
-  }
-
-  removeGoal(goal) {
-    const goalsToDestroy = [];
-
-    this.goalIds = this.goalIds.filter((gid) => {
-      const gEnt = getEntity(gid);
-
-      const isSelf = Boolean(gEnt.id === goal.id);
-      const isSiblingGoal = Boolean(
-        gEnt.goal.originalIntent &&
-          gEnt.goal.originalIntent.id === goal.goal.originalIntent.id
-      );
-
-      if (isSelf || isSiblingGoal) {
-        goalsToDestroy.push(gEnt.goal);
-        return false;
-      }
-
-      return true;
+    const evt = this.entity.fireEvent("get-bored-goals", {
+      interactor: this.entity,
+      goals: [],
     });
 
-    goalsToDestroy.forEach((gEnt) => gEnt.destroy());
-  }
+    const evts = [];
 
-  pushGoal(goal) {
-    goal.parent = this.entity;
-    return this.goalIds.push(goal.id);
-  }
+    if (this.entity.inventory) {
+      this.entity.inventory.contentIds.forEach((eid) => {
+        const ent = world.getEntity(eid);
 
-  popGoal() {
-    const eid = this.goalIds.pop().goal;
-    return getEntity(eid);
-  }
+        const evt = ent.fireEvent("get-bored-goals", {
+          interactor: this.entity,
+          goals: [],
+        });
+        if (evt) {
+          evts.push(evt);
+        }
+      });
+    }
 
-  peekGoal() {
-    // if we always have a bored goal, we don't need to check for eid...
-    const eid = this.goalIds[this.goalIds.length - 1];
-    if (eid) {
-      return getEntity(eid);
+    const goals = [...evt.data.goals, ..._.flatMap(evts, (e) => e.data.goals)];
+
+    const goal = _.sample(goals);
+
+    // console.log({ evt, goal });
+
+    if (goal) {
+      goal.source.fireEvent(goal.evt);
     }
   }
 }
