--- src/ecs/components/MoveTo.component.js
@@ -1,9 +1,0 @@
-import { Component } from "geotic";
-
-export default class MoveTo extends Component {
-  static properties = {
-    x: 0,
-    y: 0,
-    z: 0,
-  };
-}
