+++ src/ecs/index.js
@@ -17,7 +17,7 @@ import LockComponent from "./components/Lock.component";
 import LockPickComponent from "./components/LockPick.component";
 import LootComponent from "./components/Loot.component";
 import LuxComponent from "./components/Lux.component";
-import MoveToComponent from "./components/MoveTo.component";
+import MotorComponent from "./components/Motor.component";
 import PCComponent from "./components/PC.component";
 import PositionComponent from "./components/Position.component";
 import RevealedComponent from "./components/Revealed.component";
@@ -57,7 +57,7 @@ engine.registerComponent(LockComponent);
 engine.registerComponent(LockPickComponent);
 engine.registerComponent(LootComponent);
 engine.registerComponent(LuxComponent);
-engine.registerComponent(MoveToComponent);
+engine.registerComponent(MotorComponent);
 engine.registerComponent(PCComponent);
 engine.registerComponent(PositionComponent);
 engine.registerComponent(RevealedComponent);
