+++ src/ecs/prefabs/Goblin.prefab.js
@@ -14,5 +14,7 @@ export const Goblin = {
     },
     { type: "Display", properties: { name: "Goblin" } },
     { type: "Brain" },
+    { type: "Inventory" },
+    { type: "Motor" },
   ],
 };
